feat: [DP-355] add mocked unit tests
